refactor: reformat code

+++ contracts/OraclePrices.sol
@@ -7,23 +7,35 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/access/AccessControlEnumerable.sol";
 
 interface IUniswapV3Factory {
-    function getPool(address tokenA, address tokenB, uint24 fee) external view returns (address pool);
+    function getPool(
+        address tokenA,
+        address tokenB,
+        uint24 fee
+    ) external view returns (address pool);
 }
 
-interface IUniswapPool { 
-    function slot0() external view returns(uint160, int24, uint16, uint16, uint16, uint8, bool);
+interface IUniswapPool {
+    function slot0()
+        external
+        view
+        returns (
+            uint160,
+            int24,
+            uint16,
+            uint16,
+            uint16,
+            uint8,
+            bool
+        );
+
     function token0() external view returns (address);
 }
 
 interface IKeyProtocolValues {
-    function preLaunch() external pure returns(bool);
+    function preLaunch() external pure returns (bool);
 }
 
-contract OraclePrices is 
-    Context,
-    Ownable,
-    AccessControlEnumerable
-{
+contract OraclePrices is Context, Ownable, AccessControlEnumerable {
     bytes32 public constant PRICE_SETTER_ROLE = keccak256("PRICE_SETTER_ROLE");
 
     mapping(string => uint256) public prices; ///["SOY" = 560000, ...]
@@ -35,8 +47,15 @@ contract OraclePrices is
     IUniswapV3Factory public uniswapFactory;
     address public usdc = address(0x4FEB71333c2A9fE81625a5727ab0Ed33dC77B841); // to be changed for mainnet
 
-    constructor(address _address, address _xWheat, address _xSoy, address _xCorn, address _xRice, address _keyProtocolValue, address _uniswapFactory)
-    {
+    constructor(
+        address _address,
+        address _xWheat,
+        address _xSoy,
+        address _xCorn,
+        address _xRice,
+        address _keyProtocolValue,
+        address _uniswapFactory
+    ) {
         _setupRole(DEFAULT_ADMIN_ROLE, _address);
 
         prices["SOY"] = 577244585;
@@ -50,7 +69,7 @@ contract OraclePrices is
         xTokenPrices[_xRice] = 6000000;
         keyProtocolValues = IKeyProtocolValues(_keyProtocolValue);
         uniswapFactory = IUniswapV3Factory(_uniswapFactory);
-	}
+    }
 
     function setGrainPrice(string memory grain, uint256 price) public {
         require(hasRole(PRICE_SETTER_ROLE, msg.sender), "not price setter");
@@ -64,27 +83,31 @@ contract OraclePrices is
         xTokenPrices[xToken] = price;
     }
 
-    function getXTokenPrice(address xToken) public view returns(uint256) {
-        if(keyProtocolValues.preLaunch() == true) {
+    function getXTokenPrice(address xToken) public view returns (uint256) {
+        if (keyProtocolValues.preLaunch() == true) {
             return xTokenPrices[xToken];
         }
 
         address pool = getXtokenPool(xToken);
         if (pool == address(0)) {
-             return xTokenPrices[xToken];
+            return xTokenPrices[xToken];
         }
-        
+
         address poolToken0 = IUniswapPool(pool).token0();
         uint160 sqrtPriceX96;
 
-        (sqrtPriceX96,,,,,,) =  IUniswapPool(pool).slot0();
+        (sqrtPriceX96, , , , , , ) = IUniswapPool(pool).slot0();
         if (poolToken0 == usdc) {
-           return 1e12 / ((uint(sqrtPriceX96) * uint(sqrtPriceX96) * 1e6) >> (96 * 2));
+            return
+                1e12 /
+                ((uint256(sqrtPriceX96) * uint256(sqrtPriceX96) * 1e6) >>
+                    (96 * 2));
         }
-        return ((uint(sqrtPriceX96) * uint(sqrtPriceX96) * 1e6) >> (96 * 2));
+        return ((uint256(sqrtPriceX96) * uint256(sqrtPriceX96) * 1e6) >>
+            (96 * 2));
     }
 
-    function getXtokenPool(address xToken) public view returns(address) {
+    function getXtokenPool(address xToken) public view returns (address) {
         return IUniswapV3Factory(uniswapFactory).getPool(usdc, xToken, 3000);
     }
 }
